sputnik-mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sputnik-mcp"
version = "0.1.0"
description = "MCP server for Sputnik spaceship control"
requires-python = ">=3.11"
dependencies = [
    "fastmcp==2.2.1",
    "requests==2.32.3",
    "python-dotenv==1.1.0",
    "pydantic==2.11.3",
    "httpx>=0.26.0",
    "uvicorn>=0.30.0"
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "mypy",
    "pytest",
    "pytest-asyncio"
]

[tool.hatch.build.targets.wheel]
packages = ["src/sputnik_mcp"] 